requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
PyMuPDF>=1.23.7
python-dotenv>=1.0.0
httpx>=0.25.2
pillow>=10.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
loguru>=0.7.2
tenacity>=8.2.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 